Static Analysis Improvements & General Maintenance
- Removes irrelevant dot files
- Update export-ignores
- Stop ignoring platform requirements on PHP 8.1
- Update Psalm and Laminas Coding Standard to most recent versions
- Solve nearly all psalm issues (Except 2) 💪
- Various other minor improvements
