+++ .gitattributes
@@ -1,10 +1,14 @@
-/.coveralls.yml export-ignore
 /.gitattributes export-ignore
 /.github/ export-ignore
 /.gitignore export-ignore
-/.travis.yml export-ignore
 /docs/ export-ignore
 /mkdocs.yml export-ignore
 /phpcs.xml export-ignore
+/psalm.xml export-ignore
+/psalm-baseline.xml export-ignore
 /phpunit.xml.dist export-ignore
 /test/ export-ignore
+/.psr-container.php.stub
+/renovate.json
+/.laminas-ci.json
+/TODO.md
