+++ .gitignore
@@ -1,7 +1,5 @@
 /.phpcs-cache
 /.phpunit.result.cache
-/clover.xml
-/coveralls-upload.json
 /docs/html/
 /laminas-mkdoc-theme.tgz
 /laminas-mkdoc-theme/
