+++ .laminas-ci.json
@@ -1,5 +1,2 @@
 {
-    "ignore_php_platform_requirements": {
-        "8.1": true
-    }
 }
