+++ psalm-baseline.xml
@@ -1,140 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<files psalm-version="4.10.0@916b098b008f6de4543892b1e0651c1c3b92cbfa">
-  <file src="src/LazySession.php">
-    <MixedInferredReturnType occurrences="1">
-      <code>string</code>
-    </MixedInferredReturnType>
-    <MixedReturnStatement occurrences="1">
-      <code>$this-&gt;proxiedSession-&gt;getId()</code>
-    </MixedReturnStatement>
-    <UndefinedInterfaceMethod occurrences="1">
-      <code>getId</code>
-    </UndefinedInterfaceMethod>
-  </file>
+<files psalm-version="4.24.0@06dd975cb55d36af80f242561738f16c5f58264f">
   <file src="src/Session.php">
-    <MixedAssignment occurrences="1">
-      <code>$this-&gt;sessionLifetime</code>
-    </MixedAssignment>
-    <MixedInferredReturnType occurrences="1">
-      <code>null|bool|int|float|string|array|stdClass</code>
-    </MixedInferredReturnType>
+    <MixedInferredReturnType occurrences="1"/>
     <MixedReturnStatement occurrences="1">
       <code>json_decode(json_encode($value, JSON_PRESERVE_ZERO_FRACTION), true)</code>
     </MixedReturnStatement>
-  </file>
-  <file src="src/SessionMiddlewareFactory.php">
-    <MixedArgument occurrences="1">
-      <code>$container-&gt;get(SessionPersistenceInterface::class)</code>
-    </MixedArgument>
-  </file>
-  <file src="test/ConfigProviderTest.php">
-    <MixedFunctionCall occurrences="1">
-      <code>($this-&gt;provider)()</code>
-    </MixedFunctionCall>
-    <UndefinedThisPropertyAssignment occurrences="1">
-      <code>$this-&gt;provider</code>
-    </UndefinedThisPropertyAssignment>
-    <UndefinedThisPropertyFetch occurrences="1">
-      <code>$this-&gt;provider</code>
-    </UndefinedThisPropertyFetch>
-  </file>
-  <file src="test/Persistence/CacheHeadersGeneratorTraitTest.php">
-    <MixedArgument occurrences="2">
-      <code>$response-&gt;getHeaderLine('Expires')</code>
-      <code>$response-&gt;getHeaderLine('Expires')</code>
-    </MixedArgument>
-    <MixedAssignment occurrences="6">
-      <code>$response</code>
-      <code>$response</code>
-      <code>$response</code>
-      <code>$response</code>
-      <code>$response</code>
-      <code>$response</code>
-    </MixedAssignment>
-    <MixedInferredReturnType occurrences="4">
-      <code>array</code>
-      <code>array</code>
-      <code>array</code>
-      <code>iterable</code>
-    </MixedInferredReturnType>
-    <MixedMethodCall occurrences="31">
-      <code>addCacheHeadersToResponse</code>
-      <code>addCacheHeadersToResponse</code>
-      <code>addCacheHeadersToResponse</code>
-      <code>addCacheHeadersToResponse</code>
-      <code>addCacheHeadersToResponse</code>
-      <code>addCacheHeadersToResponse</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getLastModified</code>
-      <code>getLastModified</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>hasHeader</code>
-      <code>responseAlreadyHasCacheHeaders</code>
-      <code>responseAlreadyHasCacheHeaders</code>
-    </MixedMethodCall>
-    <PossiblyNullPropertyAssignmentValue occurrences="1">
-      <code>$lastModified</code>
-    </PossiblyNullPropertyAssignmentValue>
-  </file>
-  <file src="test/Persistence/SessionCookieAwareTraitTest.php">
-    <InternalMethod occurrences="1">
-      <code>trait_isDeleteCookieOnEmptySession</code>
-    </InternalMethod>
-    <MixedArgument occurrences="2">
-      <code>$actualHeaderLine</code>
-      <code>$setCookie</code>
-    </MixedArgument>
-    <MixedAssignment occurrences="9">
-      <code>$actualHeaderLine</code>
-      <code>$response</code>
-      <code>$response</code>
-      <code>$setCookie</code>
-      <code>$setCookie</code>
-      <code>$setCookie</code>
-      <code>$setCookie</code>
-      <code>$setCookie</code>
-      <code>$setCookie</code>
-    </MixedAssignment>
-    <MixedInferredReturnType occurrences="3">
-      <code>array</code>
-      <code>array</code>
-      <code>array</code>
-    </MixedInferredReturnType>
-    <MixedMethodCall occurrences="13">
-      <code>addSessionCookieHeaderToResponse</code>
-      <code>addSessionCookieHeaderToResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>createSessionCookieForResponse</code>
-      <code>getHeaderLine</code>
-      <code>getHeaderLine</code>
-      <code>getSessionCookieLifetime</code>
-      <code>getSessionCookieValueFromRequest</code>
-      <code>getSessionCookieValueFromRequest</code>
-    </MixedMethodCall>
-    <PossiblyInvalidCast occurrences="1">
-      <code>$setCookie</code>
-    </PossiblyInvalidCast>
   </file>
 </files>
