+++ test/SessionTest.php
@@ -9,7 +9,6 @@ use Mezzio\Session\SessionCookiePersistenceInterface;
 use Mezzio\Session\SessionIdentifierAwareInterface;
 use Mezzio\Session\SessionInterface;
 use PHPUnit\Framework\TestCase;
-use stdClass;
 
 use function json_decode;
 use function json_encode;
@@ -110,7 +109,7 @@ class SessionTest extends TestCase
     }
 
     /**
-     * @psalm-return array<string, array{stdClass, array<array-key, mixed>}>
+     * @psalm-return array<string, array{object, array<array-key, mixed>}>
      */
     public function serializedDataProvider(): array
     {
